pyproject.toml
[tool.poetry]
name = "analyse"
version = "0.1.0"
description = ""
authors = ["Tomas Gavenciak <gavento@gmail.com>"]
packages = [
    { include = "datalib" },
]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.2"
pandas = "^1.0.3"
xarray = "^0.15.1"
tables = ">=3.2"
python-dateutil = "^2.8.1"
countryinfo = "^0.1.0"
pycountry = "^19.8.18"
unidecode = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
black = "^19.10b0"

[tool.black]
line-length = 98
target-version = ['py37']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

